Add RoT tile attribute
The RoT tile attribute is set in the hardware (currently in gem5) for
the tile that contains the Unique Device Secret (UDS) and the Boot ROM.
It can be used e.g. in the boot.xml configuration to start the RoT
service on the same tile that booted up the Root of Trust layers.