rot: Adapt for virtual endpoints in TCU
With recent changes to the TCU the RoT needs to set up the memory region
that will be used for the virtual endpoints of the kernel. This should
be configurable and included in the signed M3 configuration report.

Add the option to the ROSA configuration structure and increase the
magic version since the binary representation changed.

The configuration only includes the number of *pages* for the EP region
because the current implementation in ROSA can only allocate full pages
in memory anyway and this saves both space and validation effort.