build: support multi-ISA builds for the same target.
Currently, this is a compromise between a fully capable multi-ISA build
and simplicity plus reasonably short build times. We could of course
build for all supported ISAs for each target, but this results in long
build times and also raises multiple other problems like which binary to
choose for boot modules or which tile/binary to choose at runtime etc.

Since we do currently not plan to have wildly mixed ISA systems, but
merely want to be able to use the RISCV-32 cores on the hardware
platform that otherwise has RISCV-64 cores, I decided to use a simpler
solution.

The idea is that we still have a default ISA for which everything is
build, but can choose a different ISA for individual applications. That
is, every application still exists just once, which simplifies a few
things.

However, this is currently only supported for Rust applications. For C++
it should be possible as well, but we need to build more libraries for
all supported ISAs, which brings up the next limitation we currently
have: all libraries, object files, and linker scripts are currently
build for all supported ISAs. That's probably fine, because they are
quick to build anyway. But it would be nicer to only build them for the
ISAs we actually need them for.