Fixed race between buf-free and credit refill.
If the receive-buffer slot is free'd after giving the credits back, the
sender could send another message before the slot is free. However,
errors on the reply-receiving side can only be caused due to
misconfiguration anyway (receive EP invalid, not enough space in receive
buffer, etc.). We therefore free the slot first and then reply to the
sender, so that a new message with the returned credit can immediately
be sent.