ci: removed test pod from k8s.sh.
We can't share this pod and it only makes the script more difficult to
use. I think it's much clearer if k8s.sh only interacts with the CI pod.
Anybody is of course free to create a new test pod, but we shouldn't
have infrastructure for that in M³ IMO.