switch from &'static Message to OwnedMessage
Previously, the kernel extensively used &'static Message for received
messages.
However, the 'static lifetime is actually wrong and the message only
lives as long as it is not answered.
To avoid mistakes and to solve possible UB, the message slot is now
managed by an RAII struct (OwnedMessage).
It prevents accessing the message after is has been answered.