Split activate syscall into multiple ones.
The activate syscall had too many responsibilities IMO, leading to a
quite long and complex implementation in the kernel. There was also only
a small amount of code shared between the activation of different gate
types and the invalidation of EPs. Furthermore, only receive gates
needed two additional arguments, requiring unnecessary additional
checks.

For that reason and also in preparation for other kernel changes, I've
now split the syscall into 4: activate_{mgate,sgate,rgate} and
invalidate. This leads already to simplifications in the kernel and
hopefully even more in the future. In userspace, we mostly know which
one to use anyway and for receive gates we needed a special case anyway.