kernel: keep Rc<Service> within ServObject.
We cannot hand out a reference from a ServObject to the inner
Rc<Service> as this would risk that someone holds that across an async
call. It's a bit different here though, because the Rc is not for a
kernel object, but an object that a kernel object refers to. That is, it
will be destroyed as soon as the last ServObject is destroyed. However,
we still have the same problem: if someone holds an Rc<Service> across
an async call it will stick around.

For that reason, I've rearranged the API of ServObject so that we do no
longer need to hand out the inner Rc<Service>.