kernel: safe way to create cap from AsyncRc.
Previously, we had a impl of From<AsyncRc<T>> for KObject, which was not
safe as you could just keep the KObject or clone the inner Rc.

To solve that, I'm now using my own conversion trait, whose method is
unsafe. This method is only used in Capability::new, which is now
generic of AsyncRc<T> and performns the conversion itself, making sure
that it directly stores the resulting KObject.