kernel: make access to syscall message safe.
If an activity is destroyed, we need to ensure that no more messages are
stored in our syscall receive buffer. This also invalidates a
potentially currently running syscall of that activity. For that reason,
we need to ensure that we do no longer access the message or reply to
the message.

We therefore store now an OwnedMessage of the current syscall in the
activity. This allows us to invalidate it during activity destruction
and also requires us to first get to the activity before trying to
reply.

The generic handler function therefore just stores the syscall message
into the current activity and each syscall function obtains a Ref to the
OwnedMessage and deserializes the syscall message based on it.

This also forces us to get rid of the message before doing async calls.
Typically we can just drop it right after decoding.