Reworked sysc_err! and moved it to libbase (fixes #14).
The sysc_err macro contained a return statement and therefore hides
control flow. Apart from that, it has actually nothing to do with
syscall calls, but is just creating a VerboseError.

I've therefore moved the macro to libbase, rename it to verror, and
changed it so that only a VerboseError is constructed.

I think it's still useful to have this macro as we use it a lot in the
kernel and besides the name being shorter, it implicitly calls format!
if multiple arguments are given.