kernel: added assertions for strong refs in release.
Now that we only keep weak references during async calls, we know the
number of references to kernel objects in revoke (mostly). That is, if
we revoke the "root" (non-derived) capability to a specific kernel
object, everything else should already be gone and therefore there
should just be one reference left.

However, that's still not true for Activity, KMem, and TileObject due to
the links between them references elsewhere.