Split derive_srv into two syscalls (fixes #22).
The problem was that we replied to the calling applications early and
then kept a thread alive that waited for the response of the server.
Therefore, applications could perform the derive_srv many times in a row
to let the kernel run out of threads.

Since derive_srv was the only syscall that had this problem, I've
decided to split this into two parts: derive_srv_req which just sends
the request to the server and derive_srv_fin which is called by the
server to finish the operation.

While simple on the first glance, it's actually quite tricky to get it
correct as we now have to keep state for that purpose at different
places: the currently running derive at Service and the information for
the completion at Activity. The split is deliberate to ensure that each
activity can just do one of these call at a time and only one derive can
be running per service. It should be possible to lift these limitations,
but it makes it even more complicated and that's currently not worth it,
I think.

Due to that state we need to ensure that it's properly removed again,
which is tricky because the service can be destroyed, the activity can
be destroyed and we cannot force the server to actually call
derive_srv_fin.