kernel: check ref count for TileObject.
The problem was that we used TileObject (and EPObject) internally in the
kernel's TileMux data structure and gave userspace access to the same
objects. However, the reason why we do that is only that we need to
remember that a MemGate for a PMP EP is stored on a particular EPObject
in case it's revoked. And that in turn requires a TileObject.

We should therefore not mix kernel-internal data structures with kernel
objects that are accessed from userspace via capabilities. Now we no
longer do, because we create a new TileObject for root so that it's no
longer referring to the same TileObject used in TileMux.

Arguably, we shouldn't even use TileObject and EPObject in
kernel-internal data structures, but changing that would be more
involved I think.