kernel: improved cap structure for Tile/KMem.
Previously, we stored unlinked caps as SEL_TILE, SEL_KMEM, and SEL_ACT
into each cap table. That is, they had no parent and thus we not
considered "derived" (wrong word here actually). That lead to a special
case when destructing them. Furthermore, we stored a strong reference of
KMem into every activity.

I've now restructured that to properly link these caps to their parents.
That is, tile and kmem are childs of these caps in the parent activity
and the parent holds the root activity for the child, whereas the child
has a child cap of it.

Furthermore, I store a weak reference of KMem into the activity and only
give back quota in Capability::release if it's upgradable. The upgrade
will only fail for root, because in all other cases the parent still has
the KMem for the child when the child is destructed.