kernel: simplify activity stop/exit.
Previously, we also revoked all capabilities when stopping an activity
and if the activity exited. That's actually not necessary and seems
weird. Thus, stop/exit now only brings the activity in a DEAD state and
cuts all connections. Revoking the activity (done by the parent) will
revoke all caps within that activity.

Furthermore, we should not store a strong reference in the activity
manager, because caused the last special case in Capability::release
(where the ref count wasn't 1 when releasing the root cap). So, now we
store a weak reference and remove activities from the activity manager
when they are dropped.

With that changed, we now have a ref count of 1 for all kernel
objects whenever the root cap is revoked.