kernel: reworked activity stop/revoke.
As it turns out, we need to revoke all caps when stopping the activity.
The reason is that otherwise it becomes basically impossible to
communicate with a child activity: let's say that parent sends a message
to the child's RecvGate to pass some work to it and waits for its
completion. If the child now crashes (and thus stops itself), the parent
would never wake up again if we do not revoke any caps. That's because
revoking the caps revokes the RecvGate and thus also the attached
SendGate and this wakes up the parent.

I don't see a reasonable way for the parent to prevent this problem.
Thus, I think stop needs to revoke capabilities as well. This is also
reasonable from a resource perspective and that's why I think other OSes
do that as well by cleaning up the process before it becomes a zombie,
for example.

Changing that led to other interesting problems, but I think we had them
in any case. We need to prevent that we stop an activity (or revoke its
caps) concurrently. That's because if one has already started to revoke
one cap, but is still waiting for it to be fully removed, another one
thinks it's already gone and falsely reports that everything is
revoked/stopped. Therefore, we now let others wait if a stop/revoke is
already in process and notify them on completion.