Reworked async Rc types.
So far, we used Rc and Weak within our custom async types, which had the
disadvantage that the memory for the RcBox, including the inner object,
was not freed until the last weak reference was removed.

For that reason and also in preparation for other features, we are now
implementing our own reference counting types. In contrast to Rc and
Weak from core, we can free the memory as soon as the last strong
reference was removed and only keep the weak references around. This
comes with an additional allocation, though.

These changes also lead to better names for our types: StrongRc, WeakRc,
and TempRc. The former two are comparable with Rc and Weak from core.
TempRc however represents a temporary reference to the object hold by a
StrongRc, which cannot be hold across an async call.