kernel: enforce uniqueness when creating cap.
To prevent that we keep other references to the kernel object elsewhere,
we now assert that there are no other references when creating the
capability (which owns the kernel object).

There are a couple of exceptions where that isn't possible and therefore
we need an unchecked constructor for the capability as well. For
example, when we cannot simply drop the kernel object on insert errors,
but need to do additional cleanup with that object.