librustbase: fix UB issue with peek_prev (fixes #40).
The problem was previously that we used the current element to get to
the previous element. Since we already handed out a mutable reference to
the current element and needed another one to get to its predecessor,
miri complained.

We now simply remember the previous element in the iterator so that we
can simply hand out a mutable reference to it when that's required.
Since that's always a different element than the current one, handing
out mutable references to both is fine. The new approach is slightly
more expensive as we need to do that even if peek_prev is not called,
but since the memory map is the only user of DList, this seems fine at
the moment.