kernel: use one additional thread per active tile (fixes #7).
That is, we now add a new thread whenever a tile is started and remove
it again if it's stopped. The kernel memory for the thread will be
accounted to the caller that started the tile, i.e., the owner of the
tile. But this is not yet done.

With that change, we also do no longer need to add a new thread to
deinit all tiles. Since the deinitialization is only required for active
tiles, we always have a thread for every such tile available that can do
the (async) deinit.