straccmux: handle exit response asynchronously.
As the kernel might come back to us with a stop sidecall after we sent
the exit call, we need to wait for the response asynchronously. This is
not perfect yet, because we might get in trouble if the kernel creates a
new activity in the meantime and this exits before the last one's exit
was ACKed, but we leave that for later.