Fix race condition when invalidating send EPs (fixes #8).
Previously, we checked first in RecvGate::receive whether the send EP
that was used to send a message to the receiver from whom we now expect
a reply. If that send EP was invalid we concluded that the receiver does
not want to talk to us anymore and thus stopped waiting for the reply.

However, as EP invalidation is only sent as a sidecall to TileMux, which
in turn will simply resume the application afterwards, this check is
racy. We could check it first, then receive the sidecall, and then do
the blocking wait-for-msg call to TileMux, which will never return as
TileMux doesn't know anything about the associated send EP.

To resolve that, the check has now been moved into TileMux which checks
it atomically. That's easy as interrupts are disabled in TileMux anyway.
For that reason, TileMux is also informed about another EP (invalidation
EP), which unblocks the activity as soon as this EP has been
invalidated.

Additionally, such an invalidation EP is also passed to the TCU's SLEEP
command, which will also wakeup the CPU as soon as this EP was
invalidated. However, this will probably remain gem5-only (to reduce the
simulation time).