Added exclusive memory regions for TEEs.
The idea is to define memory regions that only one specific tile can
access. Since TEEs will be tile-based, that's sufficient. The regions
are stored in memory tiles by a limited set of registers.

As the exclusive-region registers belong to memory tiles, we track these
resources via the tile capability. That is, userspace now also gets
access to memory tiles (not just compute tiles), but the only supported
operation is derive_tile. A tile capability additional has a
exclusive-region quota (which is 0 for compute tiles). This can be
splitted in the usual way via derive_tile.

We therefore have a new mgate_mkexcl syscall that marks this region as
exclusive. This requires that the region is power-of-2 sized and
aligned, though. The syscall requires a memory-gate capability and a
capability to the compute tile that should become the owner of the
memory region. Additionally, it requires a capability to the memory tile
from which to take the exclusive-region resource. The memory-gate needs
to belong to this memory tile.

Currently, only resource managers have access to memory tiles, because
these are the ones that create TEEs and therefore the only ones that
need this access. It's not finished, but "TEEs" can already be created
via <dom tee="1"> ... </dom> in the XML files. Setting this property
will tell the resource manager to mark the memory regions accessible by
the tile as exclusive.