Added tile_lock syscall to make EP region exclusive.
This syscall is a bit forward-looking as the actual tile locking is still
missing. However, I believe this tile lock should also mark the EPs
region as exclusive. As we are still at the exclusive-region feature,
we'll only do that now and add the rest later.

So, currently, the kernel just marks the EPs region (if existing) as
exclusive so that even the kernel cannot access it anymore. This is done
by keeping a TileObject for the memory tile that holds these EPs regions
for the kernel and using a fixed quota (currently 4) for it. The
TileObject for the userspace is derived from that and receives the rest
of the quota.