hashing: combine hashmux functionality into rot service
Current system designs more or less require that the KECACC hash accelerator
be on the same tile as the ROT hardware. To simplify things we combine them,
allowing each to run without the other if so configured.

The services are combined for opcode compactness, so 'hash' services now need
to use the 'rot' gname.