kernel: don't cause errors in critical path.
The create_map syscall tests whether the cap already exists and it's
actually the common case that it does not. Now that error creation is
actually more expensive, we need to be more careful. For that reason, we
now have a try_get method on CapTable, which just returns an Option
instead of an anyhow::Result. The same holds for try_get_mut, which how
also exists.