kernel: introduce single place where anyhow! is used.
As we rely on the root causing bying of type Error and it's pretty easy
to stick something else into the anyhow! macro, we should have some form
of protection against this. I've therefore now introduced two functions
(kerror and kerrno) which take an Error or Code and turn it into an
anyhow::Error.