Added frozen endpoints.
Endpoints in locked TCUs are now automatically frozen when they are
changed. And frozen endpoints can neither be changed, nor used, nor
invalidated and thus stay like that until they are unfrozen. This can
only be done via an unprivileged command (which can only be triggered
locally) and thus is done by the application. The application therefore
has time to validate the EP configuration before making the EP usable.

M³ uses that feature to validate receive EPs and thereby ensure that the
kernel cannot violate the integrity or confidentiality of TEEs. This is
done automatically in libm3 when activating gates. In particular, we
check the receive-buffer address and size and the reply-EP offset. The
former required a new call to TileMux to translate a virtual address to
the physical address.