ci: use Python instead of bash for the gem5 tests.
Bash got more and more annoying over time and now that we ran into the
issue with trap, which can only register one handler per signal, I've
decided to reimplement the script in Python. That gets rid of basically
all issues we had. For example, job management in Python is really easy.