Build gem5 in M³'s build directory.
Using gem5's default build directory causes trouble in our CI, because
it leads to a cached directory within a cached directory. We therefore
sidestep this problem by just building gem5 in M³'s build directory.

Arguably this is the better solution anyway, because we build everything
else in there as well.