gem5log: fixed handling of ISR exits.
Previously, we simply ignored potential returns after ISR exits, which
works most of the time, but breaks if we had a return before the ISR and
continue to the old function after the ISR.

Furthermore, we need to handle TileMux's sleep routine special here,
because that is essentially a separate thread, but we do not handle it
as one (as that would be complicated). Therefore, we have the heuristic
now that if we return from an ISR to the same binary, but a different
function, it's just an ISR exit, but no return. This way, we still
handle an ISR exit to a new ISR correctly.