gem5log: use generated thread ids.
Previously, we used the stack address to identify threads.
Unfortunately, we didn't know the stack address from the very beginning,
which led us to use '0' as a placeholder for some time.

I've resolved that now by adding generated thread ids, which are the
only ones that are user facing. Internally we still use stack addresses
and map these to the generated ids. Furthermore, we take advantage of
the fact that we only have one unknown address per binary at a time. We
therefore map '0' as a placeholder to the generated id until we found
the real one and then remove the placeholder mapping. This way, the
user-facing id stays the same all the time.