kernel: rework slab allocator
This switches from the linker-based weaving of the Rust allocator to the
new Rust-native global_allocator interface.
This brings the allocator API closer to Rust.
This commit also optimizes the slab allocator to work header-less.
Individual areas do not need a header anymore.
Instead of using the header to determine the slab of a to-be-freed
allocation, the new implementation uses the Rust-provided layout information
to determine the correct slab allocator to use.
This should hopefully reduce memory consumption with little change to
performance.