Added dynamic bit to EPs and mkdyn command.
The dynamic bit provides an exception to the frozen concept by
explicitly allowing changes to a specific endpoint even if the TCU is
locked without requiring an acknowledgement by the CU. The dynamic bit
can be freely set while the TCU is not locked, but can only be set with
the unprivileged command MKDYN otherwise. That is, while the TCU is
locked the frozen bit keeps the current value, ignoring the value from
external writes to the endpoint.

We currently use the dynamic bit for the file protocol as this requires
that the server can update the client's EP without an acknowledgement by
the client. In particular, we need this functionality for accelerators
that are not expected to deal with this. In theory, we could unfreeze
the EP after every request to the server for software clients, but this
seems unnecessary and would result in more overhead.