Added dynamic argument to alloc_ep syscall.
This allows to mark EPs as dynamic directly when allocating, in which
case the kernel will set it to dynamic. This is required in cases where
we allocate EPs for other components (e.g., for accelerators) and those
components can't mark the EP as dynamic on their own.

However, this only works if the tile is not locked yet. For
accelerators, we therefore establish all communication channels upfront
including the dynamic EP and then lock the tile. Afterwards we do not
make any further changes. For software we might want to have more
flexibility and create new communication channels even if our tile is
locked. In this case we need to be able to mark EPs as dynamic on our
own instead of asking the kernel.