Use exclusive regions for all tiles.
Previously, we only used exclusive memory regions in memory tiles. But
since we want to support accelerator TEEs as well, we also need to
protect their potentially existing tile-local memory (SPM). This can
also be helpful for normal core tiles to provide others access to their
own memory.

For that reason I've now decided that all tiles should have exclusive
regions. Potentially we have more of those for memory tiles and just a
few for compute tiles at some point, but currently the count is the same
in all TCUs.

Another change is that we now allow memory requests to locked tiles if
they hit an exclusive region. We can therefore lock an accelerator tile
and only provide specific other tiles access to specific regions of the
accelerator memory.