Added support for shared exclusive regions.
We now support overlaps of exclusive regions to use them as shared
memory for multiple tiles. To that end, each region has a locked flag
that indicates whether overlaps can still be created. As the locked bit
cannot be unset (without invalidating the region), additional tiles can
only be added before locking it.

The mgate_mkexcl syscall therefore takes a "locked" argument now that
indicates whether the region should be locked. If it is, there cannot
already be any locked regions that overlap or the syscall fails.
Additionally, all overlapping regions will be locked if the new region
is locked. If it is not locked, other overlapping regions can still be
added afterwards.