resmng: fixed bug in MemPool::make_exclusive.
We cannot simply recreate the slices that easily, because we might have
already done allocations. And if we did we lose the information about
them as we also recreate the memory map.

However, there is an easy fix: just let alloc_pool already know whether
this should be exclusive and if so derive a new MemGate for each slice
in it so that we do not have to recreate anything afterwards. So, later
when doing the make_exclusive we just to the syscalls for the individual
slices and that's it.