librustthread: replace vector-stored stack with pointer
Storing the stack as a vector is technically undefined behavior because
the compiler could assume that the contents of the vector never change as
they are not manipulated through the vector.
However, the threads actually alter the data through their stack pointers.
Thus, the stack should be stored as a pointer.

This implementation stores the stack now as a pointer to a fixed-sized
array of MaybeUninit.
This makes it safe to dereference that stack pointer in non-async
contexts; like during initialization.