resmng: store memory tile in MemMod.
Instead of searching for the memory tile later when trying to make
regions exclusive, we now store the memory tile upfront in MemMod. This
has the advantage that the MemoryManager's methods do not need access to
Resources later, which can be problematic as the MemoryManager itself is
also accessed through Resources.

Furthermore, we now allow to make slices exclusive instead of letting
the MemPool know whether it's exclusive and then making the added slices
exclusive. This enables us to not just make MemPool's exclusive, but
also individual memory slices.