root: fixed bug in receive buffer creation.
I think this worked only by accident so far, because we just took the
beginning of the MemMod for the receive buffer, which is not necessarily
the region we reserved. We therefore derive now to obtain our own cap
for the reserved region and map the beginning of this one.