resmng: added exclusive shared-memory regions.
We now can request shared-memory regions at the resource manager from
the XML file and provide specific tiles access to it. These are made
exclusive to ensure that only these tiles have access.

This is done with arguments to the resource manager and two new tags.
The arguments create the region and therefore need the name and the size
of the region: args="root shmem=myregion:64K". Afterwards, we can give
individual domains access via <shmem name="..."/> as a child of a dom
tag. To make use of the region, an application should get access to the
memory capability for that region via <shmem name="..."/> as a child of
an app tag.