rot: added rots as another boot stage.
Instead of loading it later from root, we now load it as an additional
stage after rosa. This allows us to lock the tile in order to protect
ourselfs from the kernel.

The order of events is:
1. configure sidecall EP
2. make environment exclusive region for root tile
3. lock RoT tile
4. load rots
5. start kernel
6. unimux within rots waits for sidecall
7. root starts rots on RoT tile

The sidecall EP has to be configured by rosa, because unimux uses it to
wait until rots can start. All other mux EPs and standard EPs are
created by the kernel and unfrozen (after checks) by unimux/rots before
using them.

Afterwards, rots provides root access to its environment to receive
information from root. However, it does not provide access to the full
environment, but only to the second half (unused by rots). As rots is
accessing the environment during runtime, root could change it all the
time because rots cannot revoke the access to this exclusive region. For
that reason, root loads the environment to the second part of this
region and rots just copies a few values out of it and ignores it
afterwards.

In the next step, rosa locks the tile to prevent that the kernel can
mess with anything afterwards. Now rots and the kernel are loaded and
rosa switches to the unimux part of rots.

Since root starts rots like everything else except that it doesn't load
rots into memory (it's already there), rots just works as every other M³
service using system calls and calls to its resource manager (root).
Root also writes the environment to the RoT tile to pass information to
rots. However, it ignores most of it as we do not want to trust root and
therefore have to be careful. The only important piece of information is
currently the selector for the resource manager cap.