rots: let client provide secret area.
As we want to lock the RoT tile, we can no longer store the secrets in
their and provide clients access. Instead, we now let the client
allocate memory and delegate it to rots as the secret area. That has
also the advantage that the client can chose whether this area should be
exclusive (e.g., for TEEs to protect it from the kernel).