ci: create run.sh to analyze test failures with b.
The idea is to reproduce the exact run that failed in CI, but not with
the gem5.py, but the b script. That allows us to use all debugging
facilities of b without adding all of that to the gem5.py as well.
Therefore, gem5.py is only intended to run CI tests, whereas b is used
for interactive usage.

The run.sh drops the user into a shell to get exactly the environment
the test was run in. Afterwards, b can be used to run it, debug it, add
log flags etc.