unimux: reduced heap size.
This is achieved by switching to a simple heap implementation instead of
using musl. The musl heap has quite some memory overhead, because it
allocates a significant portion of memory for every first allocation of
a differently sized type.

The simple heap only allocates the requested memory sizes and thus
allows us to shrink the bss size of rots further. However, the simple
heap does currently not support free, which is fine for the ROT, I
think.