rust: specify all crates in src/Cargo.toml.
Previously, we did not specify the library crates, which was fine
as long as they were referenced by other crates in the workspace. If
they were not (e.g., vtermcli), cargo complained that a crate was within
a workspace filesystem-wise, but wasn't listed as one.

The previous fix was to exclude these crates from the workspace, but
this had the downside that, for example, calling clippy on these crates
created a Cargo.lock file in their directory.

I think the proper way is to simply list all crates as workspace members
and thereby avoid these problems.