build: specify build.rs in Cargo.toml, not via symlink.
The main downside of a symlink is that if we add one to make a specific
crate use a build.rs, ninja considers this as "nothing has changed". And
that's correct, because without further changes we really only added the
symlink as another dependency, but following the symlink leads to a file
that indeed did not change.

Specifying the src/build.rs in the Cargo.toml of the crates that want to
use it therefore is more robust.