rot: merge rosa stages into one.
The motivation for the split into two stages was two fold:
1) we want to clear the secrets before starting rots
2) we want to prevent that the kernel can read secrets from us

The second is obsolete now, because we lock our tile and therefore
prevent anyone from reading any secrets. The first is obsolete because,
in contrast to earlier, we do another context switch to rots (i.e., we
clear the secrets), so that we still avoid the risk of rots itself
leaking any secrets later.

The split therefore does not buy us anything anymore, but makes things
more complicated, so that I've removed it now.