rot: replaced riscv-rt with our own startup.
The problem is that riscv-rt assumes instructions like csrw to be
available, which they are not on the Pico riscv32 core. As these cannot
be disabled via feature riscv-rt does more than necessary anyway I've
decided to do not use this crate anymore.

Instead, we are providing the startup by ourself, like we do already for
every other baremetal program on M³. This also means that we do no
longer copy the data to the expected location during startup (which
riscv-rt did for us). In order to clear everything between stages we
link it such that data+bss comes first, followed by text, but keep a
single jump at the start to jump over the data. In this way, we can load
the program as a single blob into memory and jump to the beginning (we
don't now the entry point as we do not have something like ELF) and at
the same time we can just zero everything up to the text. That's because
the jump can be overwritten as we don't need it anymore after the
startup.